pyproject.toml
[tool.poetry]
name = "cmd-n-ctrl"
version = "0.1.0"
description = "This applications controls my personal devices"
authors = ["igoras1993 <igor.kantorski@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uvicorn = "^0.27.1"
fastapi = "^0.109.2"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.26"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
httpx = "^0.27.0"
coloredlogs = "^15.0.1"
pytest = "^8.3.3"
email-validator = "^2.1.1"
joserfc = "^0.12.0"
pydantic = "^2.9.2"
pyplugin-lib = "^0.2.0"
psycopg2 = "^2.9.9"
textual = "^0.83.0"
textual-serve = "^1.1.1"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pre-commit = "^3.6.2"
pyright = "^1.1.356"
conventional-pre-commit = "^3.2.0"
flake8 = "^7.1.0"
Flake8-pyproject = "^1.2.3"
git-cliff = "^2.5.0"
textual-dev = "^1.6.1"



[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.27.0"
asgi-lifespan = "^2.1.0"
pytest-vcr = "^1.0.2"


[tool.poetry.group.docs.dependencies]
git-cliff = "^2.3.0"


[tool.black]
line-length = 90
preview = true
enable-unstable-feature = ["string_processing"]

[tool.flake8]
extend-ignore = ['E203', 'E701']
filename = "*.py"
exclude = [
    ".git", "__pycache__"
]
max-line-length = 90
max-complexity = 10

[tool.pyright]
reportMissingTypeStubs = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
